+++ Changelog.md
@@ -7,6 +7,8 @@ Lilu Changelog
 - Added strrchr API
 - Changed requestAccess to include API version to workaround kext loading issues
 - Updated capstone to 3.0.5 rc2
+- Improved 32-bit userspace patcher
+- Enforced `-liluslow` in installer and recovery
 
 #### v1.0
 - Initial release